requirements.txt
Flask[async]==3.0.3
gunicorn==23.0.0
httpx==0.27.2
openpyxl==3.1.5
python-docx==1.1.2
aiohttp==3.10.5
werkzeug==3.0.4
beautifulsoup4==4.12.3